Found the reason for the really poor memory usage - was reading the wrong field for the subclass.
The feature class is now extended to show all of the biotypes, and the
coverage plot also now uses all of these, but there are now too many to
be sensible.  I think we'll need to think about which ones we want to
keep, which we throw away and which we might want to merge.  Not sure
there's going to be an easy answer to this.